refactor: extract SvelteMarkdown utility helpers
